+++ Projekt_inz_backend/Controllers/DndSubclassController.cs
@@ -26,21 +26,40 @@ namespace Projekt_inz_backend.Controllers
         }
 
         [HttpGet, AllowAnonymous]
+        [ProducesResponseType(StatusCodes.Status200OK)]
         public IActionResult GetSubclasses()
         {
             return Ok(_mapper.Map<List<DndSubclassDto>>(_subclassrepos.GetSubclasses()));
         }
         [HttpGet("id/{subclassid}"), AllowAnonymous]
+        [ProducesResponseType(StatusCodes.Status200OK)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
         public IActionResult GetSubclass(int subclassid)
         {
-            return Ok(_mapper.Map<DndSubclassDto>(_subclassrepos.GetSubclass(subclassid)));
+            var subclass = _mapper.Map<DndSubclassDto>(_subclassrepos.GetSubclass(subclassid));
+            if (subclass == null)
+            {
+                return NotFound();
+            }
+            return Ok(subclass);
         }
         [HttpGet("class/{classid}"), AllowAnonymous]
+        [ProducesResponseType(StatusCodes.Status200OK)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
         public IActionResult GetSubclassesFromClass(int classid)
         {
-            return Ok(_mapper.Map<List<DndSubclassDto>>(_subclassrepos.GetSubclassesFromClass(classid)));
+            var subclasses = _mapper.Map<List<DndSubclassDto>>(_subclassrepos.GetSubclassesFromClass(classid));
+            if (subclasses == null)
+            {
+                return NotFound();
+            }
+            return Ok(subclasses);
         }
         [HttpPost, Authorize(Roles = "user,admin")]
+        [ProducesResponseType(StatusCodes.Status200OK)]
+        [ProducesResponseType(StatusCodes.Status500InternalServerError)]
+        [ProducesResponseType(StatusCodes.Status400BadRequest)]
+        [ProducesResponseType(StatusCodes.Status401Unauthorized)]
         public IActionResult CreateSubclass(int classid, [FromBody] DndSubclassDto subclass)
         {
             subclass.subclassId = null;
@@ -58,26 +77,48 @@ namespace Projekt_inz_backend.Controllers
             return Ok("Succesfuly created");
         }
         [HttpPut, Authorize(Roles = "user,admin")]
+        [ProducesResponseType(StatusCodes.Status204NoContent)]
+        [ProducesResponseType(StatusCodes.Status500InternalServerError)]
+        [ProducesResponseType(StatusCodes.Status403Forbidden)]
+        [ProducesResponseType(StatusCodes.Status401Unauthorized)]
         public IActionResult UpdateSubclass([FromBody] DndSubclassDto subclass)
         {
-            var subclassMap = _mapper.Map<DndSubclass>(subclass);
-            if (!_subclassrepos.UpdateSubclass(subclassMap))
+            if (_subclassrepos.GetOwnerId(subclass.subclassId.Value) == _subclassrepos.GetUserIdByName(_userservice.GetName())
+                || _userservice.GetRole() == "admin")
             {
-                ModelState.AddModelError("", "Cos poszlo nie tak z aktualizacja");
-                return StatusCode(500, ModelState);
+                var subclassMap = _mapper.Map<DndSubclass>(subclass);
+                if (!_subclassrepos.UpdateSubclass(subclassMap))
+                {
+                    ModelState.AddModelError("", "Cos poszlo nie tak z aktualizacja");
+                    return StatusCode(500, ModelState);
+                }
+                return NoContent();
             }
-            return NoContent();
+            ModelState.AddModelError("", "Nie masz uprawnien do zmiany tego obiektu");
+            return StatusCode(403, ModelState);
+
         }
         [HttpDelete, Authorize(Roles = "user,admin")]
+        [ProducesResponseType(StatusCodes.Status204NoContent)]
+        [ProducesResponseType(StatusCodes.Status500InternalServerError)]
+        [ProducesResponseType(StatusCodes.Status403Forbidden)]
+        [ProducesResponseType(StatusCodes.Status401Unauthorized)]
         public IActionResult DeleteSubclass([FromBody] DndSubclassDto subclass)
         {
-            var subclassMap = _mapper.Map<DndSubclass>(subclass);
-            if (!_subclassrepos.DeleteSubclass(subclassMap))
+            if (_subclassrepos.GetOwnerId(subclass.subclassId.Value) == _subclassrepos.GetUserIdByName(_userservice.GetName())
+                || _userservice.GetRole() == "admin")
             {
-                ModelState.AddModelError("", "Cos poszlo nie tak z usunieciem");
-                return StatusCode(500, ModelState);
+                var subclassMap = _mapper.Map<DndSubclass>(subclass);
+                if (!_subclassrepos.DeleteSubclass(subclassMap))
+                {
+                    ModelState.AddModelError("", "Cos poszlo nie tak z usunieciem");
+                    return StatusCode(500, ModelState);
+                }
+                return NoContent();
             }
-            return NoContent();
+            ModelState.AddModelError("", "Nie masz uprawnien do usuniecia tego obiektu");
+            return StatusCode(403, ModelState);
+
         }
     }
 }
