+++ Projekt_inz_backend/Controllers/EnemyController.cs
@@ -1,9 +1,11 @@
 ﻿using AutoMapper;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Projekt_inz_backend.Dto;
 using Projekt_inz_backend.Interfaces;
 using Projekt_inz_backend.Models;
 using Projekt_inz_backend.Repository;
+using Projekt_inz_backend.Services.UserServices;
 using System.Diagnostics;
 
 // For more information on enabling Web API for empty projects, visit https://go.microsoft.com/fwlink/?LinkID=397860
@@ -16,34 +18,55 @@ namespace Projekt_inz_backend.Controllers
     {
         private readonly IEnemyRepository _enemyrepos;
         private readonly IMapper _mapper;
+        private readonly IUserService _userservice;
 
-        public EnemyController(IEnemyRepository enemyrepos, IMapper mapper)
+        public EnemyController(IEnemyRepository enemyrepos, IMapper mapper, IUserService userservice)
         {
             _enemyrepos = enemyrepos;
             _mapper = mapper;
+            _userservice = userservice;
         }
         // GET: api/<EnemyController>
-        [HttpGet]
+        [HttpGet, AllowAnonymous]
+        [ProducesResponseType(StatusCodes.Status200OK)]
         public IActionResult GetAllEnemies()
         {
             return Ok(_mapper.Map<List<EnemyDto>>(_enemyrepos.GetEnemies()));
         }
 
         // GET api/<EnemyController>/5
-        [HttpGet("owner/{ownerid}")]
+        [HttpGet("owner/{ownerid}"), AllowAnonymous]
+        [ProducesResponseType(StatusCodes.Status200OK)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
         public IActionResult GetEnemyByOwner(int ownerid)
         {
-            return Ok(_mapper.Map<List<EnemyDto>>(_enemyrepos.GetEnemiesByOwner(ownerid)));
+            var enemies = _mapper.Map<List<EnemyDto>>(_enemyrepos.GetEnemiesByOwner(ownerid));
+            if (enemies == null)
+            {
+                return NotFound();
+            }
+            return Ok(enemies);
         }
-        [HttpGet("id/{enemyid}")]
+        [HttpGet("id/{enemyid}"), AllowAnonymous]
+        [ProducesResponseType(StatusCodes.Status200OK)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
         public IActionResult GetEnemyById(int enemyid)
         {
-            return Ok(_mapper.Map<EnemyDto>(_enemyrepos.GetEnemyById(enemyid)));
+            var enemy = _mapper.Map<EnemyDto>(_enemyrepos.GetEnemyById(enemyid));
+            if (enemy == null)
+            {
+                return NotFound();
+            }
+            return Ok(enemy);
         }
 
         // POST api/<EnemyController>
-        [HttpPost]
-        public IActionResult CreateEnemy(int ownerid, [FromBody] EnemyDto enemy)
+        [HttpPost, Authorize(Roles = "user,admin")]
+        [ProducesResponseType(StatusCodes.Status200OK)]
+        [ProducesResponseType(StatusCodes.Status500InternalServerError)]
+        [ProducesResponseType(StatusCodes.Status400BadRequest)]
+        [ProducesResponseType(StatusCodes.Status401Unauthorized)]
+        public IActionResult CreateEnemy([FromBody] EnemyDto enemy)
         {
             enemy.enemyId = null;
             if (enemy == null)
@@ -52,7 +75,7 @@ namespace Projekt_inz_backend.Controllers
             }
             var enemyMap = _mapper.Map<Enemy>(enemy);
 
-            if (!_enemyrepos.CreateEnemy(ownerid, enemyMap))
+            if (!_enemyrepos.CreateEnemy(_enemyrepos.GetUserIdByName(_userservice.GetName()), enemyMap))
             {
                 ModelState.AddModelError("", "Cos poszlo nie tak z zapisem");
                 return StatusCode(500, ModelState);
@@ -61,29 +84,51 @@ namespace Projekt_inz_backend.Controllers
         }
 
         // PUT api/<EnemyController>/5
-        [HttpPut]
+        [HttpPut, Authorize(Roles = "user,admin")]
+        [ProducesResponseType(StatusCodes.Status204NoContent)]
+        [ProducesResponseType(StatusCodes.Status500InternalServerError)]
+        [ProducesResponseType(StatusCodes.Status401Unauthorized)]
+        [ProducesResponseType(StatusCodes.Status403Forbidden)]
         public IActionResult UpdateEnemy(EnemyDto enemy)
         {
-            var enemyMap = _mapper.Map<Enemy>(enemy);
-            if (!_enemyrepos.UpdateEnemy(enemyMap))
+            if (_enemyrepos.GetOwnerId(enemy.enemyId.Value) == _enemyrepos.GetUserIdByName(_userservice.GetName())
+                || _userservice.GetRole() == "admin")
             {
-                ModelState.AddModelError("", "Cos poszlo nie tak z aktualizacja");
-                return StatusCode(500, ModelState);
+                var enemyMap = _mapper.Map<Enemy>(enemy);
+                if (!_enemyrepos.UpdateEnemy(enemyMap))
+                {
+                    ModelState.AddModelError("", "Cos poszlo nie tak z aktualizacja");
+                    return StatusCode(500, ModelState);
+                }
+                return NoContent();
             }
-            return NoContent();
+            ModelState.AddModelError("", "Nie masz uprawnien do zmiany tego obiektu");
+            return StatusCode(403, ModelState);
+
         }
 
         // DELETE api/<EnemyController>/5
-        [HttpDelete]
+        [HttpDelete, Authorize(Roles = "user,admin")]
+        [ProducesResponseType(StatusCodes.Status204NoContent)]
+        [ProducesResponseType(StatusCodes.Status500InternalServerError)]
+        [ProducesResponseType(StatusCodes.Status401Unauthorized)]
+        [ProducesResponseType(StatusCodes.Status403Forbidden)]
         public IActionResult Delete(EnemyDto enemy)
         {
-            var enemyMap = _mapper.Map<Enemy>(enemy);
-            if (!_enemyrepos.DeleteEnemy(enemyMap))
+            if (_enemyrepos.GetOwnerId(enemy.enemyId.Value) == _enemyrepos.GetUserIdByName(_userservice.GetName())
+                || _userservice.GetRole() == "admin")
             {
-                ModelState.AddModelError("", "Cos poszlo nie tak z usunieciem");
-                return StatusCode(500, ModelState);
+                var enemyMap = _mapper.Map<Enemy>(enemy);
+                if (!_enemyrepos.DeleteEnemy(enemyMap))
+                {
+                    ModelState.AddModelError("", "Cos poszlo nie tak z usunieciem");
+                    return StatusCode(500, ModelState);
+                }
+                return NoContent();
             }
-            return NoContent();
+            ModelState.AddModelError("", "Nie masz uprawnien do usuniecia tego obiektu");
+            return StatusCode(403, ModelState);
+
         }
     }
 }
